gossip/temporary package for tmp dbs
